+++ src/main/java/com/pockit/pockit/Repositories/TransactionRepository.java
@@ -3,5 +3,5 @@ package com.pockit.pockit.Repositories;
 import com.pockit.pockit.Models.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction,Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 }
